Refactor codec libraries and introduce type_t
Move the encoding algorithms from functions in coded_pkg to procedures
in codec_builder_pkg. This brings symmetry to the pacakage definitions.
Change the encode and decode procedures to handle only the data for
array types. Decoding the array range is now handled entirely by the
functions in codec_pkg.

Introduce type_t and integrate it with both the codec and queue
packages. Provide a code_length function that takes a type_t and an
array length (if any) to determine the expected length for an encoded
string of that type.

Place type_t into a new package types_pkg. Move range_t there as well.
Together they form a basis for future work on a generic type system.