Update ext_ptr_pkg with better GHDL VHPIDIRECT types
Update ext_ptr_pkg to make better use of the available types in the GHDL
VHPIDIRECT interface. Notably, provide functions to return "views" of
the external data as a string access or integer_vector access type. The
data can be manipulated through these views, but the access view should
not be deallocated.