Store ext_ptr as pointer rather than access_t
Previously, ext_ptr shared memory was structured to mimic the structure
of GHDL VHPIDIRECT access types, so that the memory could be directly
manipulated in VHDL. After more thought, the use cases for this are
minimal. Instead, switch back to storing a simple, bare pointer. Provide
VHDL functions to return the entire shared memory block as a string or
integer_vector.